test analytics: fix upload

+++ misc/python/materialize/test_analytics/data/feature_benchmark/feature_benchmark_result_storage.py
@@ -72,7 +72,7 @@ class FeatureBenchmarkResultStorage(BaseDataStorage):
                     {as_sanitized_literal(result_entry.scale)},
                     {result_entry.is_regression},
                     {result_entry.wallclock.result or 'NULL::DOUBLE'},
-                    NULL,  -- to be removed when we remove the "messages" column
+                    NULL::INT,  -- to be removed when we remove the "messages" column
                     {result_entry.memory_mz.result or 'NULL::DOUBLE'},
                     {result_entry.memory_clusterd.result or 'NULL::DOUBLE'},
                     {result_entry.wallclock.min or 'NULL::DOUBLE'},
@@ -121,7 +121,7 @@ class FeatureBenchmarkResultStorage(BaseDataStorage):
                     {discarded_entry.cycle},
                     {discarded_entry.is_regression},
                     {discarded_entry.wallclock.result or 'NULL::DOUBLE'},
-                    NULL,  -- to be removed when we remove the "messages" column
+                    NULL::INT,  -- to be removed when we remove the "messages" column
                     {discarded_entry.memory_mz.result or 'NULL::DOUBLE'},
                     {discarded_entry.memory_clusterd.result or 'NULL::DOUBLE'},
                     {discarded_entry.wallclock.min or 'NULL::DOUBLE'},
